backend/requirements.txt
# ── Core Backend ──────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
psycopg[binary]==3.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
alembic==1.13.1

# ── Data Processing ──────────────────────────────────
numpy==1.26.4

# ── Computer Vision ──────────────────────────────────
opencv-python==4.10.0.84
mediapipe==0.10.14

# ── Face Recognition (install separately on Windows) ─
# Step 1: pip install dlib (see install_dlib.md)
# Step 2: pip install face-recognition==1.3.0
